+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.5.3] - 2024-09-11
+
 ### Added
 
 - Filter `ubb_possible_links_filter_sql` to add to the query for possible post links.
@@ -493,7 +495,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 First Release!
 
-[unreleased]: https://github.com/26b/unbabble/compare/0.5.2...HEAD
+[unreleased]: https://github.com/26b/unbabble/compare/0.5.3...HEAD
+[0.5.3]: https://github.com/26b/unbabble/compare/0.5.2...0.5.3
 [0.5.2]: https://github.com/26b/unbabble/compare/0.5.1...0.5.2
 [0.5.1]: https://github.com/26b/unbabble/compare/0.5.0...0.5.1
 [0.5.0]: https://github.com/26b/unbabble/compare/0.4.8...0.5.0